Organize component structure
