+++ src/components/FirecallItems/DeleteFirecallItemDialog.tsx
@@ -1,7 +1,7 @@
 import { useCallback } from 'react';
 import { useFirecallId } from '../../hooks/useFirecall';
 import useFirecallItemUpdate from '../../hooks/useFirecallItemUpdate';
-import ConfirmDialog from '../ConfirmDialog';
+import ConfirmDialog from '../dialogs/ConfirmDialog';
 import { FirecallItem } from '../firebase/firestore';
 import { firecallItems } from './infos/firecallitems';
 import { FirecallItemInfo } from './infos/types';
