+++ src/components/FirecallItems/FirecallItemDialog.tsx
@@ -11,8 +11,8 @@ import Select, { SelectChangeEvent } from '@mui/material/Select';
 import TextField from '@mui/material/TextField';
 import moment from 'moment';
 import React, { useState } from 'react';
-import ConfirmDialog from '../ConfirmDialog';
-import MyDateTimePicker from '../DateTimePicker';
+import ConfirmDialog from '../dialogs/ConfirmDialog';
+import MyDateTimePicker from '../inputs/DateTimePicker';
 import { FirecallItem } from '../firebase/firestore';
 import { firecallItemInfo, firecallItems } from './infos/firecallitems';
 import { FirecallItemInfo } from './infos/types';
