+++ src/pages/_app.tsx
@@ -2,11 +2,11 @@ import CssBaseline from '@mui/material/CssBaseline';
 import type { AppProps } from 'next/app';
 import Head from 'next/head';
 import React from 'react';
-import AppDrawer from '../components/AppDrawer';
+import AppDrawer from '../components/site/AppDrawer';
 import FirebaseUserProvider from '../components/firebase/FirebaseUserProvider';
-import FirecallProvider from '../components/FirecallProvider';
-import HeaderBar from '../components/HeaderBar';
-import SingedOutOneTapLogin from '../components/SingedOutOneTapLogin';
+import FirecallProvider from '../components/providers/FirecallProvider';
+import HeaderBar from '../components/site/HeaderBar';
+import SingedOutOneTapLogin from '../components/auth/SingedOutOneTapLogin';
 import '../styles/globals.css';
 import styles from '../styles/Home.module.css';
