+++ src/pages/tokens.tsx
@@ -3,7 +3,7 @@ import dynamic from 'next/dynamic';
 
 const DynamicTokens = dynamic(
   () => {
-    return import('../components/Tokens');
+    return import('../components/pages/Tokens');
   },
   { ssr: false }
 );
